Create check_ha_portal_svr_dist.py
When you have a high availability (HA) environment, it implies having two Portal for ArcGIS servers installed and configured in the environment (primary and secondary server). 

Through potal REST, you can validate which of the two servers is the primary and which is the secondary, thus allowing you to identify them and once you know that you can for example apply patches on the secondary node without affecting the primary.

This python script uses the ArcGIS API created by ESRI, so that you can enter as a parameter url of the portal, the name of a user and password (an administrator user); with this the script makes the connection to the REST of Portal for ArcGIS, validate which are the Potal servers that are part of the environment and retona as a response the name of each of the servers and if it is the primary server, or the secondary.